requirements/base.txt
Django==4.1
django-cors-headers==3.13.0
django-stubs==1.12.0
django-stubs-ext==0.5.0
djangorestframework==3.13.1
djangorestframework-stubs==1.7.0
mypy==0.971
mypy-extensions==0.4.3
psycopg2-binary==2.9.3
python-dotenv==0.20.0
martor==1.6.13
Pillow==9.2.0
django-storages==1.13.1
boto3==1.24.51
celery==5.2.7
redis==4.3.4
celery-stubs==0.1.2